PortalNetwork: Remove static constructors in PortalNetwork and UltralightProvider (#757)
* Remove static function in Portalnetwork and create new constructor file

* Remove instances of the static method in test files

* Migrate to ethjs release candidates (#754)

* remove node-gyp

* update ethjs deps

* migrate portalnetwork to ethjs breaking release

* fix integration tests

* migrate cli to use breaking releases

* update package lock

* add typescript back to portalnetwork deps

* package-lock

* Remove static constructor from test files

* Remove more static constructor instances

* fix imports and revert test change

---------
