+++ packages/portalnetwork/src/client/index.ts
@@ -3,3 +3,4 @@ export * from './dbManager.js'
 export * from './provider.js'
 export * from './routingTable.js'
 export * from './types.js'
+export { createPortalNetwork } from './constructor.js'
