+++ packages/portalnetwork/test/client/client.spec.ts
@@ -1,9 +1,9 @@
 import { assert, describe, it } from 'vitest'
 
-import { NetworkId, PortalNetwork, TransportLayer } from '../../src/index.js'
+import { NetworkId, createPortalNetwork, TransportLayer } from '../../src/index.js'
 describe('Client unit tests', () => {
   it('node initialization/startup', async () => {
-    const node = await PortalNetwork.create({
+    const node = await createPortalNetwork({
       bindAddress: '192.168.0.1',
       transport: TransportLayer.WEB,
       supportedNetworks: [{ networkId: NetworkId.HistoryNetwork }],
