+++ packages/portalnetwork/test/client/provider.spec.ts
@@ -4,11 +4,10 @@ import { keys } from '@libp2p/crypto'
 import { multiaddr } from '@multiformats/multiaddr'
 import { assert, describe, expect, it } from 'vitest'
 
-import { bytesToUnprefixedHex } from '@ethereumjs/util'
-import { hexToBytes } from 'ethereum-cryptography/utils'
-import { UltralightProvider } from '../../src/client/provider.js'
+import { createUltralightProvider } from '../../src/client/provider.js'
 import { TransportLayer } from '../../src/index.js'
 import { NetworkId } from '../../src/networks/types.js'
+import { bytesToUnprefixedHex, hexToBytes } from '@ethereumjs/util'
 
 describe('Test provider functionality', () => {
   it('should test provider API', async () => {
@@ -16,7 +15,7 @@ describe('Test provider functionality', () => {
     const privateKey = await keys.generateKeyPair('secp256k1')
     const enr = SignableENR.createFromPrivateKey(privateKey)
     enr.setLocationMultiaddr(ma)
-    const provider = await UltralightProvider.create({
+    const provider = await createUltralightProvider({
       bindAddress: '0.0.0.0.0',
       transport: TransportLayer.NODE,
       config: {
@@ -135,7 +134,7 @@ describe('Test provider functionality', () => {
     await provider.portal.stop()
   })
   it('should instantiate provider with default network settings', async () => {
-    const provider = await UltralightProvider.create({})
+    const provider = await createUltralightProvider({})
     assert.ok(provider.portal.bootnodes.length > 0)
     assert.ok(provider.portal.discv5.bindAddrs[0].toString().includes('0.0.0.0'))
   })
