+++ packages/portalnetwork/test/integration/blacklist.spec.ts
@@ -4,7 +4,7 @@ import { multiaddr } from "@multiformats/multiaddr"
 import { hexToBytes } from "ethereum-cryptography/utils"
 import { assert, describe, it } from "vitest"
 import type { HistoryNetwork } from "../../src";
-import { NetworkId, PortalNetwork, TransportLayer } from "../../src/index.js"
+import { NetworkId, TransportLayer, createPortalNetwork } from "../../src/index.js"
 const privateKeys = [
     '0x0a2700250802122102273097673a2948af93317235d2f02ad9cf3b79a34eeb37720c5f19e09f11783c12250802122102273097673a2948af93317235d2f02ad9cf3b79a34eeb37720c5f19e09f11783c1a2408021220aae0fff4ac28fdcdf14ee8ecb591c7f1bc78651206d86afe16479a63d9cb73bd',
     '0x0a27002508021221039909a8a7e81dbdc867480f0eeb7468189d1e7a1dd7ee8a13ee486c8cbd743764122508021221039909a8a7e81dbdc867480f0eeb7468189d1e7a1dd7ee8a13ee486c8cbd7437641a2408021220c6eb3ae347433e8cfe7a0a195cc17fc8afcd478b9fb74be56d13bccc67813130',
@@ -19,7 +19,7 @@ describe('black list test', async () => {
   enr1.setLocationMultiaddr(initMa)
   const initMa2: any = multiaddr(`/ip4/127.0.0.1/udp/5031`)
   enr2.setLocationMultiaddr(initMa2)
-  const node1 = await PortalNetwork.create({
+  const node1 = await createPortalNetwork({
     transport: TransportLayer.NODE,
     supportedNetworks: [{ networkId: NetworkId.HistoryNetwork }],
     config: {
@@ -30,7 +30,7 @@ describe('black list test', async () => {
       privateKey: pk1,
     },
   })
-  const node2 = await PortalNetwork.create({
+  const node2 = await createPortalNetwork({
     transport: TransportLayer.NODE,
     supportedNetworks: [{ networkId: NetworkId.HistoryNetwork }],
     config: {
