+++ packages/portalnetwork/test/integration/recursiveContentLookup.spec.ts
@@ -5,7 +5,7 @@ import { keys } from '@libp2p/crypto'
 import { multiaddr } from '@multiformats/multiaddr'
 import { assert, beforeAll, describe, it } from 'vitest'
 
-import { ContentLookup, NetworkId, PortalNetwork, TransportLayer } from '../../src/index.js'
+import { ContentLookup, NetworkId, PortalNetwork, TransportLayer, createPortalNetwork } from '../../src/index.js'
 
 import type { ContentTrace, HistoryNetwork } from '../../src/index.js'
 
@@ -25,7 +25,7 @@ describe('Recursive Content Lookup Test', () => {
     enr2.setLocationMultiaddr(initMa2)
     const initMa3: any = multiaddr(`/ip4/127.0.0.1/udp/5002`)
     enr3.setLocationMultiaddr(initMa3)
-    node1 = await PortalNetwork.create({
+    node1 = await createPortalNetwork({
       transport: TransportLayer.NODE,
       supportedNetworks: [{ networkId: NetworkId.HistoryNetwork }],
       config: {
@@ -36,7 +36,7 @@ describe('Recursive Content Lookup Test', () => {
         privateKey: pk1,
       },
     })
-    node2 = await PortalNetwork.create({
+    node2 = await createPortalNetwork({
       transport: TransportLayer.NODE,
       supportedNetworks: [{ networkId: NetworkId.HistoryNetwork }],
       config: {
@@ -47,7 +47,7 @@ describe('Recursive Content Lookup Test', () => {
         privateKey: pk2,
       },
     })
-    node3 = await PortalNetwork.create({
+    node3 = await createPortalNetwork({
       transport: TransportLayer.NODE,
       supportedNetworks: [{ networkId: NetworkId.HistoryNetwork }],
       config: {
