+++ packages/portalnetwork/test/integration/state.spec.ts
@@ -9,8 +9,8 @@ import {
   AccountTrieNodeOffer,
   AccountTrieNodeRetrieval,
   NetworkId,
-  PortalNetwork,
   TransportLayer,
+  createPortalNetwork,
 } from '../../src/index.js'
 import samples from '../networks/state/testdata/accountNodeSamples.json'
 
@@ -38,7 +38,7 @@ describe('AccountTrieNode Gossip / Request', async () => {
   enr1.setLocationMultiaddr(initMa)
   const initMa2: any = multiaddr(`/ip4/127.0.0.1/udp/3021`)
   enr2.setLocationMultiaddr(initMa2)
-  const node1 = await PortalNetwork.create({
+  const node1 = await createPortalNetwork({
     transport: TransportLayer.NODE,
     supportedNetworks: [{ networkId: NetworkId.StateNetwork }],
     config: {
@@ -49,7 +49,7 @@ describe('AccountTrieNode Gossip / Request', async () => {
       privateKey: pk1,
     },
   })
-  const node2 = await PortalNetwork.create({
+  const node2 = await createPortalNetwork({
     transport: TransportLayer.NODE,
     supportedNetworks: [{ networkId: NetworkId.StateNetwork }],
     config: {
@@ -114,7 +114,7 @@ describe('getAccount via network', async () => {
       const enr = SignableENR.createFromPrivateKey(pk)
       const initMa: any = multiaddr(`/ip4/127.0.0.1/udp/${3022 + i}`)
       enr.setLocationMultiaddr(initMa)
-      const node = await PortalNetwork.create({
+      const node = await createPortalNetwork({
         transport: TransportLayer.NODE,
         supportedNetworks: [{ networkId: NetworkId.StateNetwork }],
         config: {
