+++ packages/portalnetwork/test/networks/history/blockIndex.spec.ts
@@ -5,7 +5,7 @@ import {
   BlockHeaderWithProof,
   HistoryNetworkContentType,
   NetworkId,
-  PortalNetwork,
+  createPortalNetwork,
   getContentKey,
 } from '../../../src/index.js'
 
@@ -22,7 +22,7 @@ const header = createBlockHeaderFromRLP(serializedHeader.header, { setHardfork: 
 const hash = bytesToHex(header.hash())
 
 describe('BlockIndex', async () => {
-  const ultralight = await PortalNetwork.create({
+  const ultralight = await createPortalNetwork({
     bindAddress: '127.0.0.1',
     supportedNetworks: [{ networkId: NetworkId.HistoryNetwork }],
   })
@@ -45,7 +45,7 @@ describe('BlockIndex', async () => {
     assert.deepEqual(stored, expected)
   })
 
-  const ultralight2 = await PortalNetwork.create({
+  const ultralight2 = await createPortalNetwork({
     bindAddress: '127.0.0.1',
     supportedNetworks: [{ networkId: NetworkId.HistoryNetwork }],
     db: ultralight.db.db,
